Rename functions and add comments for clarity
Move brute force algorithms out of script05 and create a new file
for it. Clean up script05 so that it imports the functions from
a new file HistogramAlgTools.py. Add comments to scripts for
clarity.
